fix merge master compile problem

+++ libraries/chain/wasm_price.cpp
@@ -520,7 +520,7 @@ namespace contento { namespace chain { namespace wasm_price {
     // contract_balance_api
     WASM_PRICE_FN_CONSTS(
                          (get_contract_balance,      (void*, int) )
-                         (transfer,                  (void*,int64_t, int))
+                         (transfer,                  (void*,int, int))
                          (get_value,                 (int64_t))
                          (pay_prohibited,            (void*))
                          );
